Move dentry resolver to its own package
