+++ pkg/security/resolvers/dentry/dentry_resolver_test.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package resolvers
+package dentry
 
 import (
 	"testing"
