+++ pkg/security/tests/dentry_test.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/security/ebpf/kernel"
 	"github.com/DataDog/datadog-agent/pkg/security/metrics"
-	"github.com/DataDog/datadog-agent/pkg/security/probe/resolvers"
+	"github.com/DataDog/datadog-agent/pkg/security/resolvers/dentry"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/rules"
 )
@@ -216,7 +216,7 @@ func BenchmarkERPCDentryResolutionSegment(b *testing.B) {
 	}
 
 	// create a new dentry resolver to avoid concurrent map access errors
-	resolver, err := resolvers.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
+	resolver, err := dentry.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -285,7 +285,7 @@ func BenchmarkERPCDentryResolutionPath(b *testing.B) {
 	}
 
 	// create a new dentry resolver to avoid concurrent map access errors
-	resolver, err := resolvers.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
+	resolver, err := dentry.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -354,7 +354,7 @@ func BenchmarkMapDentryResolutionSegment(b *testing.B) {
 	}
 
 	// create a new dentry resolver to avoid concurrent map access errors
-	resolver, err := resolvers.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
+	resolver, err := dentry.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
 	if err != nil {
 		b.Fatal(err)
 	}
@@ -423,7 +423,7 @@ func BenchmarkMapDentryResolutionPath(b *testing.B) {
 	}
 
 	// create a new dentry resolver to avoid concurrent map access errors
-	resolver, err := resolvers.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
+	resolver, err := dentry.NewDentryResolver(test.probe.Config, test.probe.StatsdClient, test.probe.Erpc)
 	if err != nil {
 		b.Fatal(err)
 	}
