+++ pkg/security/tests/schemas.go
@@ -18,7 +18,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/security/events"
 	"github.com/DataDog/datadog-agent/pkg/security/probe"
-	"github.com/DataDog/datadog-agent/pkg/security/probe/resolvers"
+	"github.com/DataDog/datadog-agent/pkg/security/resolvers/dentry"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
 )
 
@@ -50,7 +50,7 @@ func (v ValidInodeFormatChecker) IsFormat(input interface{}) bool {
 	default:
 		return false
 	}
-	return !resolvers.IsFakeInode(inode)
+	return !dentry.IsFakeInode(inode)
 }
 
 //nolint:deadcode,unused
